@@ -1,10 +1,8 @@
 ﻿using AutoMapper.QueryableExtensions;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
-using System.Data;
 using WesternStatesWater.WestDaat.Accessors.EntityFramework;
 using WesternStatesWater.WestDaat.Accessors.Mapping;
-using WesternStatesWater.WestDaat.Common;
 using WesternStatesWater.WestDaat.Common.DataContracts;
 
 namespace WesternStatesWater.WestDaat.Accessors
@@ -20,7 +18,7 @@ namespace WesternStatesWater.WestDaat.Accessors
 
         async Task<Site> ISiteAccessor.GetSiteByUuid(string siteUuid)
         {
-            using var db = _databaseContextFactory.Create();
+            await using var db = _databaseContextFactory.Create();
             return await db.SitesDim
                 .Where(x => x.SiteUuid == siteUuid)
                 .ProjectTo<Site>(DtoMapper.Configuration)
@@ -29,8 +27,7 @@ namespace WesternStatesWater.WestDaat.Accessors
 
         async Task<SiteDigest> ISiteAccessor.GetSiteDigestByUuid(string siteUuid)
         {
-            using var db = _databaseContextFactory.Create();
-            
+            await using var db = _databaseContextFactory.Create();
             return await db.SitesDim
                 .Where(x => x.SiteUuid == siteUuid)
                 .ProjectTo<SiteDigest>(DtoMapper.Configuration)
@@ -48,7 +45,7 @@ namespace WesternStatesWater.WestDaat.Accessors
 
         public async Task<SiteDetails> GetSiteDetailsByUuid(string siteUuid)
         {
-            using var db = _databaseContextFactory.Create();
+            await using var db = _databaseContextFactory.Create();
             return await db.SitesDim
                 .Where(x => x.SiteUuid == siteUuid)
                 .ProjectTo<SiteDetails>(DtoMapper.Configuration)
@@ -57,7 +54,7 @@ namespace WesternStatesWater.WestDaat.Accessors
 
         async Task<SiteLocation> ISiteAccessor.GetWaterSiteLocationByUuid(string siteUuid)
         {
-            using var db = _databaseContextFactory.Create();
+            await using var db = _databaseContextFactory.Create();
             return await db.SitesDim.Where(x => x.SiteUuid == siteUuid)
                 .ProjectTo<SiteLocation>(DtoMapper.Configuration)
                 .SingleAsync();
@@ -65,18 +62,17 @@ namespace WesternStatesWater.WestDaat.Accessors
 
         public async Task<List<WaterSourceInfoListItem>> GetWaterSiteSourceInfoListByUuid(string siteUuid)
         {
-            using var db = _databaseContextFactory.Create();
-
+            await using var db = _databaseContextFactory.Create();
             return await db.SitesDim.Where(x => x.SiteUuid == siteUuid)
-                    .SelectMany(x => x.WaterSourceBridgeSitesFact
+                .SelectMany(x => x.WaterSourceBridgeSitesFact
                     .Select(a => a.WaterSource))
-                    .ProjectTo<WaterSourceInfoListItem>(DtoMapper.Configuration)
-                    .ToListAsync();
+                .ProjectTo<WaterSourceInfoListItem>(DtoMapper.Configuration)
+                .ToListAsync();
         }
 
         public async Task<List<WaterRightInfoListItem>> GetWaterRightInfoListByUuid(string siteUuid)
         {
-            using var db = _databaseContextFactory.Create();
+            await using var db = _databaseContextFactory.Create();
             return await db.AllocationAmountsFact
                 .Where(x => x.AllocationBridgeSitesFact.Any(y => y.Site.SiteUuid == siteUuid))
                 .ProjectTo<WaterRightInfoListItem>(DtoMapper.Configuration)
@@ -95,10 +91,7 @@ namespace WesternStatesWater.WestDaat.Accessors
                 SiteName = a.SiteName,
             });
 
-            foreach (var geoConnexItem in query)
-            {
-                yield return geoConnexItem;
-            }
+            return query;
         }
     }
 }
