@@ -1,7 +1,6 @@
 ﻿using AutoMapper.QueryableExtensions;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
-using System.Data;
 using WesternStatesWater.WestDaat.Accessors.EntityFramework;
 using WesternStatesWater.WestDaat.Accessors.Mapping;
 using WesternStatesWater.WestDaat.Common.DataContracts;
@@ -19,8 +18,7 @@ namespace WesternStatesWater.WestDaat.Accessors
 
         async Task<List<BeneficialUseItem>> ISystemAccessor.GetAvailableBeneficialUseNormalizedNames()
         {
-            using var db = _databaseContextFactory.Create();
-
+            await using var db = _databaseContextFactory.Create();
             return await db.BeneficialUsesCV
                 .ProjectTo<BeneficialUseItem>(DtoMapper.Configuration)
                 .Distinct()
@@ -32,32 +30,32 @@ namespace WesternStatesWater.WestDaat.Accessors
 
         async Task<List<string>> ISystemAccessor.GetAvailableWaterSourceTypeNormalizedNames()
         {
-            using var db = _databaseContextFactory.Create();
+            await using var db = _databaseContextFactory.Create();
             return await db.WaterSourceType
-              .Select(a => a.WaDEName.Length > 0 ? a.WaDEName : a.Name)
-              .Distinct()
-              .OrderBy(a => a)
-              .ToListAsync();
+                .Select(a => a.WaDEName.Length > 0 ? a.WaDEName : a.Name)
+                .Distinct()
+                .OrderBy(a => a)
+                .ToListAsync();
         }
 
         async Task<List<string>> ISystemAccessor.GetAvailableOwnerClassificationNormalizedNames()
         {
-            using var db = _databaseContextFactory.Create();
+            await using var db = _databaseContextFactory.Create();
             return await db.OwnerClassificationCv
-              .Select(a => a.WaDEName.Length > 0 ? a.WaDEName : a.Name)
-              .Distinct()
-              .OrderBy(a => a)
-              .ToListAsync();
+                .Select(a => a.WaDEName.Length > 0 ? a.WaDEName : a.Name)
+                .Distinct()
+                .OrderBy(a => a)
+                .ToListAsync();
         }
 
         async Task<List<string>> ISystemAccessor.GetAvailableStateNormalizedNames()
         {
-            using var db = _databaseContextFactory.Create();
+            await using var db = _databaseContextFactory.Create();
             return await db.State
-              .Select(a => a.WaDEName.Length > 0 ? a.WaDEName : a.Name)
-              .Distinct()
-              .OrderBy(a => a)
-              .ToListAsync();
+                .Select(a => a.WaDEName.Length > 0 ? a.WaDEName : a.Name)
+                .Distinct()
+                .OrderBy(a => a)
+                .ToListAsync();
         }
     }
 }
