@@ -1,26 +1,20 @@
-using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.Azure.WebJobs;
-using Microsoft.Azure.WebJobs.Extensions.Http;
 using Microsoft.Extensions.Logging;
 using Newtonsoft.Json;
 using System.IO;
+using System.Net;
+using Azure.Core.Serialization;
 using WesternStatesWater.WestDaat.Contracts.Client;
+using Microsoft.Azure.Functions.Worker;
+using Microsoft.Azure.Functions.Worker.Http;
 
 namespace WesternStatesWater.WestDaat.Client.Functions
 {
-    public class Function1
+    public class Function1(ITestManager testManager) : FunctionBase
     {
-        private readonly ITestManager _testManager;
-
-        public Function1(ITestManager testManager)
-        {
-            _testManager = testManager;
-        }
-
-        [FunctionName("TestMe")]
-        public async Task<IActionResult> Run(
-            [HttpTrigger(AuthorizationLevel.Function, "get", "post", Route = null)] HttpRequest req,
+        [Function("TestMe")]
+        public async Task<HttpResponseData> Run(
+            [HttpTrigger(AuthorizationLevel.Function, "get", "post", Route = null)]
+            HttpRequestData req,
             ILogger log)
         {
             log.LogInformation("C# HTTP trigger function processed a request.");
@@ -31,9 +25,10 @@ namespace WesternStatesWater.WestDaat.Client.Functions
             dynamic data = JsonConvert.DeserializeObject(requestBody);
             name = name ?? data?.name;
 
-            var result = _testManager.TestMe("Test Me");
-
-            return new OkObjectResult(result);
+            var result = testManager.TestMe("Test Me");
+            var response = req.CreateResponse(HttpStatusCode.OK);
+            await response.WriteAsJsonAsync(result, new JsonObjectSerializer());
+            return data;
         }
     }
 }
