@@ -1,29 +1,79 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using System.Net;
 using System.Text.Json;
 using System.Text.Json.Serialization;
+using Azure.Core.Serialization;
+using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.Azure.Functions.Worker.Http;
+using Microsoft.Extensions.Logging;
 
 namespace WesternStatesWater.WestDaat.Client.Functions
 {
     public class FunctionBase
     {
-        private static readonly JsonSerializerOptions _jsonSerializerOptions = CreateJsonSerializerOptions();
-
-        public IActionResult JsonResult(object obj)
+        private static JsonSerializerOptions JsonSerializerOptions => new()
         {
-            var jsonToReturn = JsonSerializer.Serialize(obj, _jsonSerializerOptions);
-            return new ContentResult
-            {
-                Content = jsonToReturn,
-                ContentType = "application/json",
-                StatusCode = 200
-            };
+            PropertyNamingPolicy = JsonNamingPolicy.CamelCase,
+            Converters = { new JsonStringEnumConverter() }
+        };
+
+        protected static async Task<HttpResponseData> CreateOkResponse<T>(
+            HttpRequestData request,
+            T response)
+        {
+            var data = request.CreateResponse(HttpStatusCode.OK);
+
+            await data.WriteAsJsonAsync((object) response, new JsonObjectSerializer(JsonSerializerOptions));
+
+            return data;
         }
 
-        private static JsonSerializerOptions CreateJsonSerializerOptions()
+        protected static Task<HttpResponseData> CreateBadRequestResponse(HttpRequestData request, ValidationError error)
         {
-            var opts = new JsonSerializerOptions();
-            opts.Converters.Add(new JsonStringEnumConverter(JsonNamingPolicy.CamelCase));
-            return opts;
+            var details = new HttpValidationProblemDetails(error.Errors)
+            {
+                Status = (int) HttpStatusCode.BadRequest,
+                Title = "One or more validation errors occurred.",
+                Type = "https://tools.ietf.org/html/rfc7231#section-6.5.1"
+            };
+
+            return CreateProblemDetailsResponse(request, details, HttpStatusCode.BadRequest);
+        }
+
+        private static async Task<HttpResponseData> CreateProblemDetailsResponse(
+            HttpRequestData request,
+            ProblemDetails details,
+            HttpStatusCode statusCode)
+        {
+            var response = request.CreateResponse();
+
+            // Casting to object for polymorphic serialization
+            await response.WriteAsJsonAsync<object>(
+                details,
+                new JsonObjectSerializer(JsonSerializerOptions),
+                statusCode);
+
+            return response;
+        }
+
+        protected static async Task<T> Deserialize<T>(HttpRequestData request, ILogger logger) where T : class
+        {
+            T result = null;
+
+            try
+            {
+                // Workaround since every function currently calls deserialize regardless of http method type.
+                result = request.Method != "GET"
+                    ? await JsonSerializer.DeserializeAsync<T>(request.Body, JsonSerializerOptions)
+                    : null;
+            }
+            catch (JsonException ex)
+            {
+                logger.LogError(ex, "Failed to deserialize type '{TypeName}'", typeof(T).Name);
+            }
+
+            // For legacy reasons, return null instead of an error response. We should change this to return an error response.
+            return result;
         }
     }
 }
