@@ -1,19 +1,18 @@
-﻿using Microsoft.AspNetCore.Authorization;
-using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.Azure.WebJobs;
-using Microsoft.Azure.WebJobs.Extensions.Http;
-using Microsoft.Extensions.Logging;
+﻿using Microsoft.Extensions.Logging;
 using Newtonsoft.Json;
 using System.IO;
+using System.Net;
+using System.Text.Json;
+using Azure.Core.Serialization;
 using WesternStatesWater.WestDaat.Contracts.Client;
-using JsonSerializer = System.Text.Json.JsonSerializer;
+using Microsoft.Azure.Functions.Worker;
+using Microsoft.Azure.Functions.Worker.Http;
 
 namespace WesternStatesWater.WestDaat.Client.Functions
 {
     public class System : FunctionBase
     {
-        public System(ISystemManager systemManager, INotificationManager notificationManager, ILogger<WaterAllocation> logger)
+        public System(ISystemManager systemManager, INotificationManager notificationManager, ILogger<System> logger)
         {
             _systemManager = systemManager;
             _notificationManager = notificationManager;
@@ -24,82 +23,92 @@ namespace WesternStatesWater.WestDaat.Client.Functions
         private readonly ISystemManager _systemManager;
         private readonly ILogger _logger;
 
-        [FunctionName(nameof(GetBeneficialUses))]
-        public async Task<IActionResult> GetBeneficialUses([HttpTrigger(AuthorizationLevel.Anonymous, "get", Route = "system/beneficialuses")] HttpRequest req)
+        [Function(nameof(GetBeneficialUses))]
+        public async Task<HttpResponseData> GetBeneficialUses([HttpTrigger(AuthorizationLevel.Anonymous, "get", Route = "system/beneficialuses")] HttpRequestData req)
         {
             _logger.LogInformation("Getting Beneficial Uses");
 
             var results = await _systemManager.GetAvailableBeneficialUseNormalizedNames();
 
-            return new OkObjectResult(results);
+            // This cannot run thru CreateOKResponse because the serializer options will convert enum to string
+            // This will break the UI.
+            var data = req.CreateResponse(HttpStatusCode.OK);
+            await data.WriteAsJsonAsync<object>(results, new JsonObjectSerializer(
+            new JsonSerializerOptions
+            {
+                PropertyNamingPolicy = JsonNamingPolicy.CamelCase
+            }));
+            return data;
         }
 
-        [FunctionName(nameof(GetOwnerClassifications)), AllowAnonymous]
-        public async Task<IActionResult> GetOwnerClassifications([HttpTrigger(AuthorizationLevel.Anonymous, "get", Route = "system/ownerclassifications")] HttpRequest request)
+        [Function(nameof(GetOwnerClassifications))]
+        public async Task<HttpResponseData> GetOwnerClassifications([HttpTrigger(AuthorizationLevel.Anonymous, "get", Route = "system/ownerclassifications")] HttpRequestData request)
         {
             _logger.LogInformation("Getting Owner Classifications");
 
             var results = await _systemManager.GetAvailableOwnerClassificationNormalizedNames();
 
-            return JsonResult(results);
+            return await CreateOkResponse(request, results);
         }
 
-        [FunctionName(nameof(GetWaterSourceTypes)), AllowAnonymous]
-        public async Task<IActionResult> GetWaterSourceTypes([HttpTrigger(AuthorizationLevel.Anonymous, "get", Route = "system/watersourcetypes")] HttpRequest request)
+        [Function(nameof(GetWaterSourceTypes))]
+        public async Task<HttpResponseData> GetWaterSourceTypes([HttpTrigger(AuthorizationLevel.Anonymous, "get", Route = "system/watersourcetypes")] HttpRequestData request)
         {
             _logger.LogInformation("Getting Water Source Types");
 
             var results = await _systemManager.GetAvailableWaterSourceTypeNormalizedNames();
 
-            return JsonResult(results);
+            return await CreateOkResponse(request, results);
         }
 
-        [FunctionName(nameof(GetStates)), AllowAnonymous]
-        public async Task<IActionResult> GetStates([HttpTrigger(AuthorizationLevel.Anonymous, "get", Route = "system/states")] HttpRequest request)
+        [Function(nameof(GetStates))]
+        public async Task<HttpResponseData> GetStates([HttpTrigger(AuthorizationLevel.Anonymous, "get", Route = "system/states")] HttpRequestData request)
         {
             _logger.LogInformation("Getting States");
 
             var results = await _systemManager.GetAvailableStateNormalizedNames();
 
-            return JsonResult(results);
+            return await CreateOkResponse(request, results);
         }
 
-        [FunctionName(nameof(GetRiverBasinNames)), AllowAnonymous]
-        public IActionResult GetRiverBasinNames([HttpTrigger(AuthorizationLevel.Anonymous, "get", Route = "system/RiverBasins")] HttpRequest request)
+        [Function(nameof(GetRiverBasinNames))]
+        public async Task<HttpResponseData> GetRiverBasinNames([HttpTrigger(AuthorizationLevel.Anonymous, "get", Route = "system/RiverBasins")] HttpRequestData request)
         {
             var result = _systemManager.GetRiverBasinNames();
 
-            return new OkObjectResult(result);
+            return await CreateOkResponse(request, result);
         }
 
-        [FunctionName(nameof(GetRiverBasinPolygonsByName)), AllowAnonymous]
-        public async Task<IActionResult> GetRiverBasinPolygonsByName([HttpTrigger(AuthorizationLevel.Anonymous, "post", Route = "system/RiverBasins")] HttpRequest request)
+        [Function(nameof(GetRiverBasinPolygonsByName))]
+        public async Task<HttpResponseData> GetRiverBasinPolygonsByName([HttpTrigger(AuthorizationLevel.Anonymous, "post", Route = "system/RiverBasins")] HttpRequestData request)
         {
             string requestBody = String.Empty;
             using (StreamReader streamReader = new StreamReader(request.Body))
             {
                 requestBody = await streamReader.ReadToEndAsync();
             }
+            
             var basinNames = JsonConvert.DeserializeObject<string[]>(requestBody);
 
             var result = _systemManager.GetRiverBasinPolygonsByName(basinNames);
 
-            return new OkObjectResult(JsonSerializer.Serialize(result));
+            return await CreateOkResponse(request, result);
         }
 
-        [FunctionName(nameof(PostFeedback)), AllowAnonymous]
-        public async Task<IActionResult> PostFeedback([HttpTrigger(AuthorizationLevel.Anonymous, "post", Route = "system/feedback")] HttpRequest request)
+        [Function(nameof(PostFeedback))]
+        public async Task<HttpResponseData> PostFeedback([HttpTrigger(AuthorizationLevel.Anonymous, "post", Route = "system/feedback")] HttpRequestData request)
         {
             string requestBody = string.Empty;
             using (StreamReader streamReader = new StreamReader(request.Body))
             {
                 requestBody = await streamReader.ReadToEndAsync();
             }
+            
             var feedbackRequest = JsonConvert.DeserializeObject<FeedbackRequest>(requestBody);
 
             await _notificationManager.SendFeedback(feedbackRequest);
 
-            return new OkResult();
+            return request.CreateResponse(HttpStatusCode.OK);
         }
     }
 }
