@@ -1,16 +1,16 @@
-using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.Azure.WebJobs;
-using Microsoft.Azure.WebJobs.Extensions.Http;
 using Microsoft.Extensions.Logging;
 using Newtonsoft.Json;
 using System.IO;
+using System.Net;
 using WesternStatesWater.WestDaat.Common;
 using WesternStatesWater.WestDaat.Contracts.Client;
 using JsonSerializer = System.Text.Json.JsonSerializer;
 using Microsoft.AspNetCore.Http.Features;
+using Microsoft.AspNetCore.Mvc;
 using WesternStatesWater.WestDaat.Common.Exceptions;
+using Microsoft.Azure.Functions.Worker;
+using Microsoft.Azure.Functions.Worker.Http;
 
 namespace WesternStatesWater.WestDaat.Client.Functions
 {
@@ -25,8 +25,8 @@ namespace WesternStatesWater.WestDaat.Client.Functions
         private readonly IWaterAllocationManager _waterAllocationManager;
         private readonly ILogger _logger;
 
-        [FunctionName(nameof(NldiFeatures))]
-        public async Task<IActionResult> NldiFeatures([HttpTrigger(AuthorizationLevel.Anonymous, "get", Route = "NldiFeatures/@{latitude},{longitude}")] HttpRequest req, double latitude, double longitude)
+        [Function(nameof(NldiFeatures))]
+        public async Task<HttpResponseData> NldiFeatures([HttpTrigger(AuthorizationLevel.Anonymous, "get", Route = "NldiFeatures/@{latitude},{longitude}")] HttpRequestData req, double latitude, double longitude)
         {
             _logger.LogInformation("Getting NLDI Features []");
 
@@ -35,59 +35,60 @@ namespace WesternStatesWater.WestDaat.Client.Functions
 
             var results = await _waterAllocationManager.GetNldiFeatures(latitude, longitude, directions, dataPoints);
 
-            return JsonResult(results);
+            return await CreateOkResponse(req, results);
         }
 
         // Water Right Routes
-        [FunctionName(nameof(FindWaterRights)), AllowAnonymous]
-        public async Task<IActionResult> FindWaterRights([HttpTrigger(AuthorizationLevel.Anonymous, "post", Route = "WaterRights/find")] HttpRequest request)
+        [Function(nameof(FindWaterRights))]
+        public async Task<HttpResponseData> FindWaterRights([HttpTrigger(AuthorizationLevel.Anonymous, "post", Route = "WaterRights/find")] HttpRequestData request)
         {
             string requestBody = string.Empty;
             using (StreamReader streamReader = new StreamReader(request.Body))
             {
                 requestBody = await streamReader.ReadToEndAsync();
             }
+            
             var searchRequest = JsonConvert.DeserializeObject<WaterRightsSearchCriteriaWithPaging>(requestBody);
 
             var result = await _waterAllocationManager.FindWaterRights(searchRequest);
 
-            return new OkObjectResult(result);
+            return await CreateOkResponse(request, result);
         }
 
-        [FunctionName(nameof(GetWaterRightDetails)), AllowAnonymous]
-        public async Task<IActionResult> GetWaterRightDetails([HttpTrigger(AuthorizationLevel.Anonymous, "get", Route = "WaterRights/{allocationUuid}")] HttpRequest request, string allocationUuid)
+        [Function(nameof(GetWaterRightDetails))]
+        public async Task<HttpResponseData> GetWaterRightDetails([HttpTrigger(AuthorizationLevel.Anonymous, "get", Route = "WaterRights/{allocationUuid}")] HttpRequestData request, string allocationUuid)
         {
             var result = await _waterAllocationManager.GetWaterRightDetails(allocationUuid);
 
-            return new OkObjectResult(result);
+            return await CreateOkResponse(request, result);
         }
 
-        [FunctionName(nameof(GetWaterRightSiteInfoList)), AllowAnonymous]
-        public async Task<IActionResult> GetWaterRightSiteInfoList([HttpTrigger(AuthorizationLevel.Anonymous, "get", Route = "WaterRights/{allocationUuid}/Sites")] HttpRequest request, string allocationUuid)
+        [Function(nameof(GetWaterRightSiteInfoList))]
+        public async Task<HttpResponseData> GetWaterRightSiteInfoList([HttpTrigger(AuthorizationLevel.Anonymous, "get", Route = "WaterRights/{allocationUuid}/Sites")] HttpRequestData request, string allocationUuid)
         {
             var result = await _waterAllocationManager.GetWaterRightSiteInfoList(allocationUuid);
 
-            return new OkObjectResult(result);
+            return await CreateOkResponse(request, result);
         }
 
-        [FunctionName(nameof(GetWaterRightSourceInfoList)), AllowAnonymous]
-        public async Task<IActionResult> GetWaterRightSourceInfoList([HttpTrigger(AuthorizationLevel.Anonymous, "get", Route = "WaterRights/{allocationUuid}/Sources")] HttpRequest request, string allocationUuid)
+        [Function(nameof(GetWaterRightSourceInfoList))]
+        public async Task<HttpResponseData> GetWaterRightSourceInfoList([HttpTrigger(AuthorizationLevel.Anonymous, "get", Route = "WaterRights/{allocationUuid}/Sources")] HttpRequestData request, string allocationUuid)
         {
             var result = await _waterAllocationManager.GetWaterRightSourceInfoList(allocationUuid);
 
-            return new OkObjectResult(result);
+            return await CreateOkResponse(request, result);
         }
 
-        [FunctionName(nameof(GetWaterRightSiteLocations)), AllowAnonymous]
-        public async Task<IActionResult> GetWaterRightSiteLocations([HttpTrigger(AuthorizationLevel.Anonymous, "get", Route = "WaterRights/{allocationUuid}/SiteLocations")] HttpRequest request, string allocationUuid)
+        [Function(nameof(GetWaterRightSiteLocations))]
+        public async Task<HttpResponseData> GetWaterRightSiteLocations([HttpTrigger(AuthorizationLevel.Anonymous, "get", Route = "WaterRights/{allocationUuid}/SiteLocations")] HttpRequestData request, string allocationUuid)
         {
             var result = await _waterAllocationManager.GetWaterRightSiteLocations(allocationUuid);
 
-            return new OkObjectResult(JsonSerializer.Serialize(result));
+            return await CreateOkResponse(request, result);
         }
 
-        [FunctionName(nameof(GetAnalyticsSummaryInformation)), AllowAnonymous]
-        public async Task<IActionResult> GetAnalyticsSummaryInformation([HttpTrigger(AuthorizationLevel.Anonymous, "post", Route = "WaterRights/AnalyticsSummaryInformation")] HttpRequest request)
+        [Function(nameof(GetAnalyticsSummaryInformation))]
+        public async Task<HttpResponseData> GetAnalyticsSummaryInformation([HttpTrigger(AuthorizationLevel.Anonymous, "post", Route = "WaterRights/AnalyticsSummaryInformation")] HttpRequestData request)
         {
             string requestBody = string.Empty;
             using (StreamReader streamReader = new StreamReader(request.Body))
@@ -98,11 +99,11 @@ namespace WesternStatesWater.WestDaat.Client.Functions
 
             var result = await _waterAllocationManager.GetAnalyticsSummaryInformation(searchRequest);
 
-            return new OkObjectResult(result);
+            return await CreateOkResponse(request, result);
         }
 
-        [FunctionName(nameof(GetWaterRightsEnvelope)), AllowAnonymous]
-        public async Task<IActionResult> GetWaterRightsEnvelope([HttpTrigger(AuthorizationLevel.Anonymous, "post", Route = "WaterRights/DataEnvelope")] HttpRequest request)
+        [Function(nameof(GetWaterRightsEnvelope))]
+        public async Task<HttpResponseData> GetWaterRightsEnvelope([HttpTrigger(AuthorizationLevel.Anonymous, "post", Route = "WaterRights/DataEnvelope")] HttpRequestData request)
         {
             string requestBody = string.Empty;
             using (StreamReader streamReader = new StreamReader(request.Body))
@@ -113,69 +114,69 @@ namespace WesternStatesWater.WestDaat.Client.Functions
 
             var result = await _waterAllocationManager.GetWaterRightsEnvelope(searchRequest);
 
-            return new OkObjectResult(JsonSerializer.Serialize(result));
+            return await CreateOkResponse(request, result);
         }
 
         // Site Routes
-        [FunctionName(nameof(GetWaterAllocationSiteDetails)), AllowAnonymous]
-        public async Task<IActionResult> GetWaterAllocationSiteDetails([HttpTrigger(AuthorizationLevel.Anonymous, "get", Route = "sites/{siteUuid}/geoconnex")] HttpRequest request, string siteUuid)
+        [Function(nameof(GetWaterAllocationSiteDetails))]
+        public async Task<HttpResponseData> GetWaterAllocationSiteDetails([HttpTrigger(AuthorizationLevel.Anonymous, "get", Route = "sites/{siteUuid}/geoconnex")] HttpRequestData request, string siteUuid)
         {
             var result = await _waterAllocationManager.GetWaterAllocationSiteGeoconnexIntegrationData(siteUuid);
 
-            return new OkObjectResult(result);
+            return await CreateOkResponse(request, result);
         }
 
-        [FunctionName(nameof(GetSiteDetails)), AllowAnonymous]
-        public async Task<IActionResult> GetSiteDetails([HttpTrigger(AuthorizationLevel.Anonymous, "get", Route = "Sites/{siteUuid}")] HttpRequest request, string siteUuid)
+        [Function(nameof(GetSiteDetails))]
+        public async Task<HttpResponseData> GetSiteDetails([HttpTrigger(AuthorizationLevel.Anonymous, "get", Route = "Sites/{siteUuid}")] HttpRequestData request, string siteUuid)
         {
             var result = await _waterAllocationManager.GetSiteDetails(siteUuid);
 
-            return new OkObjectResult(result);
+            return await CreateOkResponse(request, result);
         }
 
-        [FunctionName(nameof(GetWaterRightSiteDigest)), AllowAnonymous]
-        public async Task<IActionResult> GetWaterRightSiteDigest([HttpTrigger(AuthorizationLevel.Anonymous, "get", Route = "Sites/{siteUuid}/WaterRightsDigest")] HttpRequest request, string siteUuid)
+        [Function(nameof(GetWaterRightSiteDigest))]
+        public async Task<HttpResponseData> GetWaterRightSiteDigest([HttpTrigger(AuthorizationLevel.Anonymous, "get", Route = "Sites/{siteUuid}/WaterRightsDigest")] HttpRequestData request, string siteUuid)
         {
             var result = await _waterAllocationManager.GetWaterRightsDigestsBySite(siteUuid);
 
-            return new OkObjectResult(result);
+            return await CreateOkResponse(request, result);
         }
 
-        [FunctionName(nameof(GetSiteDigest)), AllowAnonymous]
-        public async Task<IActionResult> GetSiteDigest([HttpTrigger(AuthorizationLevel.Anonymous, "get", Route = "Sites/{siteUuid}/Digest")] HttpRequest request, string siteUuid)
+        [Function(nameof(GetSiteDigest))]
+        public async Task<HttpResponseData> GetSiteDigest([HttpTrigger(AuthorizationLevel.Anonymous, "get", Route = "Sites/{siteUuid}/Digest")] HttpRequestData request, string siteUuid)
         {
             var result = await _waterAllocationManager.GetSiteDigest(siteUuid);
 
-            return new OkObjectResult(result);
+            return await CreateOkResponse(request, result);
         }
 
-        [FunctionName(nameof(GetSiteLocation)), AllowAnonymous]
-        public async Task<IActionResult> GetSiteLocation([HttpTrigger(AuthorizationLevel.Anonymous, "get", Route = "Sites/{siteUuid}/SiteLocation")] HttpRequest request, string siteUuid)
+        [Function(nameof(GetSiteLocation))]
+        public async Task<HttpResponseData> GetSiteLocation([HttpTrigger(AuthorizationLevel.Anonymous, "get", Route = "Sites/{siteUuid}/SiteLocation")] HttpRequestData request, string siteUuid)
         {
             var result = await _waterAllocationManager.GetWaterSiteLocation(siteUuid);
 
-            return new OkObjectResult(JsonSerializer.Serialize(result));
+            return await CreateOkResponse(request, result);
         }
 
-        [FunctionName(nameof(GetWaterSiteSourceListByUuid)), AllowAnonymous]
-        public async Task<IActionResult> GetWaterSiteSourceListByUuid([HttpTrigger(AuthorizationLevel.Anonymous, "get", Route = "Sites/{siteUuid}/Sources")] HttpRequest request, string siteUuid)
+        [Function(nameof(GetWaterSiteSourceListByUuid))]
+        public async Task<HttpResponseData> GetWaterSiteSourceListByUuid([HttpTrigger(AuthorizationLevel.Anonymous, "get", Route = "Sites/{siteUuid}/Sources")] HttpRequestData request, string siteUuid)
         {
             var result = await _waterAllocationManager.GetWaterSiteSourceInfoListByUuid(siteUuid);
 
-            return new OkObjectResult(result);
+            return await CreateOkResponse(request, result);
         }
 
-        [FunctionName(nameof(GetWaterSiteRightsListByUuid)), AllowAnonymous]
-        public async Task<IActionResult> GetWaterSiteRightsListByUuid([HttpTrigger(AuthorizationLevel.Anonymous, "get", Route = "Sites/{siteUuid}/Rights")] HttpRequest request, string siteUuid)
+        [Function(nameof(GetWaterSiteRightsListByUuid))]
+        public async Task<HttpResponseData> GetWaterSiteRightsListByUuid([HttpTrigger(AuthorizationLevel.Anonymous, "get", Route = "Sites/{siteUuid}/Rights")] HttpRequestData request, string siteUuid)
         {
             var result = await _waterAllocationManager.GetWaterSiteRightsInfoListByUuid(siteUuid);
 
-            return new OkObjectResult(result);
+            return await CreateOkResponse(request, result);
         }
 
 
-        [FunctionName(nameof(DownloadWaterRights)), AllowAnonymous]
-        public async Task<EmptyResult> DownloadWaterRights([HttpTrigger(AuthorizationLevel.Anonymous, "post", Route = "WaterRights/download")] HttpRequest request)
+        [Function(nameof(DownloadWaterRights))]
+        public async Task<IActionResult> DownloadWaterRights([HttpTrigger(AuthorizationLevel.Anonymous, "post", Route = "WaterRights/download")] HttpRequest request)
         {
             // the IO operations have to be synchronous 
             var feature = request.HttpContext.Features.Get<IHttpBodyControlFeature>();
@@ -184,26 +185,27 @@ namespace WesternStatesWater.WestDaat.Client.Functions
             using var streamReader = new StreamReader(request.Body);
             var requestBody = await streamReader.ReadToEndAsync();
             var searchRequest = JsonConvert.DeserializeObject<WaterRightsSearchCriteriaWithFilterUrl>(requestBody);
-
-            var response = request.HttpContext.Response;
-            response.Headers.Add("Content-Type", "application/octet-stream");
-            response.Headers.Append("Content-Disposition", "attachment; filename=\"WaterRights.zip\"");
-
+            var ms = new MemoryStream();
             try
             {
-                await _waterAllocationManager.WaterRightsAsZip(response.Body, searchRequest);
+                await _waterAllocationManager.WaterRightsAsZip(ms, searchRequest);
             }
-            catch (WestDaatException)
+            catch (WestDaatException wex)
             {
-                response.StatusCode = StatusCodes.Status413PayloadTooLarge;
+                request.HttpContext.Response.StatusCode = (int)HttpStatusCode.RequestEntityTooLarge;
             }
-            catch (Exception)
+            catch (Exception ex)
             {
-                response.StatusCode = StatusCodes.Status500InternalServerError;
+                request.HttpContext.Response.StatusCode = (int)HttpStatusCode.InternalServerError;
                 throw;
             }
 
-            return new EmptyResult();
+            request.HttpContext.Response.Headers.Append("Access-Control-Expose-Headers", "Content-Disposition");
+            ms.Seek(0, SeekOrigin.Begin);
+            return new FileStreamResult(ms, "application/octet-stream")
+            {
+                FileDownloadName = "WaterRights.zip"
+            };
         }
     }
 }
