@@ -1,10 +1,12 @@
-﻿using Microsoft.Extensions.Configuration;
+﻿using System.Reflection;
+using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 using WesternStatesWater.WestDaat.Common.Configuration;
 
 namespace WesternStatesWater.WestDaat.Tests.ManagerTests
 {
+    [TestCategory("Manager Tests")]
     public abstract class ManagerTestBase
     {
         private ILoggerFactory _loggerFactory;
@@ -16,6 +18,7 @@ namespace WesternStatesWater.WestDaat.Tests.ManagerTests
                                         .AddInMemoryCollection(DefaultTestConfiguration)
                                         .AddJsonFile("appsettings.json", optional: true, reloadOnChange: true)
                                         .AddJsonFile("personal.settings.json", optional: true, reloadOnChange: true)
+                                        .AddUserSecrets(Assembly.GetExecutingAssembly(), true)
                                         .AddEnvironmentVariables()
                                         .Build();
         }
